Make 'killall adb' run only if needed to avoid premature script exiting

+++ manufacturing_scripts/aopp-flash-flounder.sh
@@ -61,8 +61,11 @@ f_run(){
     exit 1
   fi
 
-  # Kill running server
-  killall adb &> /dev/null
+  # Kill server if one is already running
+  if [[ -n "$(pgrep adb)" ]]; then
+    echo "Killing server"
+    killall adb &> /dev/null
+  fi
 
   # Start server
   echo
